Move Auth client to pkg
