+++ pkg/auth/client.go
